% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mean_sd.R
\name{analyze_mean_sd}
\alias{analyze_mean_sd}
\title{Calculate mean and sd}
\usage{
analyze_mean_sd(df)
}
\arguments{
\item{df}{main dataframe for core pangenome analysis}
}
\value{
data.frame
}
\description{
Calculate mean and sd
}
