% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_median_IQR.R
\name{analyze_median_IQR}
\alias{analyze_median_IQR}
\title{Calculate median and IQR}
\usage{
analyze_median_IQR(df)
}
\arguments{
\item{df}{main dataframe for core pangenome analysis}
}
\value{
data.frame
}
\description{
Calculate median and IQR
}
