% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bool_multiplicates.R
\name{bool_multiplicates}
\alias{bool_multiplicates}
\title{Transform to boolean number of proteins per cluster to 1 if duplicates occur.}
\usage{
bool_multiplicates(x)
}
\arguments{
\item{x}{integer vector}
}
\description{
Transform to boolean number of proteins per cluster to 1 if duplicates occur.
}
