% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_set_pangenome.R
\name{calc_set_pangenome}
\alias{calc_set_pangenome}
\title{Calculates pangenome for one iteration. Returns integer vector with 
accumulative pangenome sizes.}
\usage{
calc_set_pangenome(m, comb)
}
\arguments{
\item{m}{0/1 matrix}

\item{comb}{combination vector}
}
\description{
Calculates pangenome for one iteration. Returns integer vector with 
accumulative pangenome sizes.
}
