% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{function formatting data for plotting}
\usage{
format_data(df)
}
\arguments{
\item{df}{data.frame output Get_Core_Pangenome_List.py}
}
\description{
function formatting data for plotting
}
