% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_df.R
\name{format_df}
\alias{format_df}
\title{Change structure of main table}
\usage{
format_df(df)
}
\arguments{
\item{df}{dataframe output of modify_colnames}
}
\value{
dataframe used in subsequent core pangenome analysis
}
\description{
Change structure of main table
}
