% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_core.R
\name{get_core}
\alias{get_core}
\title{get core clusters between two genomes}
\usage{
get_core(x, y)
}
\arguments{
\item{x}{integer vector genome 1}

\item{y}{integer vector genome 2}
}
\description{
get core clusters between two genomes
}
