% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res.R
\name{get_res}
\alias{get_res}
\title{gathers output for two genomes}
\usage{
get_res(x, y, core)
}
\arguments{
\item{x}{output of get_unique_X}

\item{y}{output of get_unique_Y}

\item{core}{output of get_core}
}
\description{
gathers output for two genomes
}
