% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_x.R
\name{get_unique_X}
\alias{get_unique_X}
\title{get unique clusters for genome X}
\usage{
get_unique_X(x, y)
}
\arguments{
\item{x}{integer vector genome 1}

\item{y}{integer vector genome 2}
}
\description{
get unique clusters for genome X
}
