% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_y.R
\name{get_unique_Y}
\alias{get_unique_Y}
\title{get unique clusters for genome Y}
\usage{
get_unique_Y(x, y)
}
\arguments{
\item{x}{integer vector genome 1}

\item{y}{integer vector genome 2}
}
\description{
get unique clusters for genome Y
}
