% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_df.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{converts 0/1 matrix to data.frame}
\usage{
matrix_to_df(my_bool_matrix, times = 10000)
}
\arguments{
\item{my_bool_matrix}{m0/1 matrix table}
}
\description{
converts 0/1 matrix to data.frame
}
