% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_iterations.R
\name{merge_iterations}
\alias{merge_iterations}
\title{wrapper function calculating pangenome for all iterations}
\usage{
merge_iterations(m, comb)
}
\arguments{
\item{m}{0/1 matrix}

\item{comb}{combination vector}
}
\value{
a list of iterations
}
\description{
wrapper function calculating pangenome for all iterations
}
