% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_colnames.R
\name{modify_colnames}
\alias{modify_colnames}
\title{Modify column names}
\usage{
modify_colnames(df)
}
\arguments{
\item{df}{data.frame, output of multmerge function}
}
\value{
data.frame
}
\description{
Modify column names to more readable format
}
