% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multmerge.R
\name{multmerge}
\alias{multmerge}
\title{Load multiple .csv and merge}
\usage{
multmerge(mypath)
}
\arguments{
\item{mypath}{path to folder with .csv files generated by Calc_Core_Clust_Perm.py}
}
\value{
data.frame
}
\description{
Loads all .csv file from specified location that has number before suffix.
}
