% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core_boxplot.R
\name{plot_core_boxplot}
\alias{plot_core_boxplot}
\title{function plotting publication core pangenome plot}
\usage{
plot_core_boxplot(df_clean)
}
\arguments{
\item{df_clean}{clean data.frame}
}
\description{
function plotting publication core pangenome plot
}
