% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_sd.R
\name{plot_mean_sd}
\alias{plot_mean_sd}
\title{Plot accumulation plot
Prot accumulation plot for core pangenome based on mead and standard deviation}
\usage{
plot_mean_sd(sum_df_mean)
}
\arguments{
\item{sum_df_median}{summarized table}
}
\description{
Plot accumulation plot
Prot accumulation plot for core pangenome based on mead and standard deviation
}
