% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_median_IQR.R
\name{plot_median_IQR}
\alias{plot_median_IQR}
\title{Plot accumulation plot
Prot accumulation plot for core pangenome based on median and IQR}
\usage{
plot_median_IQR(sum_df_median)
}
\arguments{
\item{sum_df_median}{summarized table}
}
\description{
Plot accumulation plot
Prot accumulation plot for core pangenome based on median and IQR
}
