% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pangenome_boxplt.R
\name{plot_pangenome_boxplt}
\alias{plot_pangenome_boxplt}
\title{plotting accumulation plot for pangenome}
\usage{
plot_pangenome_boxplt(out_df)
}
\arguments{
\item{out_df}{final data.frame}
}
\description{
plotting accumulation plot for pangenome
}
