% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_merged_iterations.R
\name{tidy_merged_iterations}
\alias{tidy_merged_iterations}
\title{clean up data.frame}
\usage{
tidy_merged_iterations(out)
}
\arguments{
\item{out}{data.frame}
}
\description{
clean up data.frame
}
